+++ aiosmtpd/proxy_protocol.py
@@ -261,7 +261,8 @@ class ProxyData:
     """
     Represents data received during PROXY Protocol Handshake, in an already-parsed form
     """
-
+    
+    # pytype: disable=annotation-type-mismatch
     version: Optional[int] = attr.ib(kw_only=True, init=True)
     """PROXY Protocol version; None if not recognized/malformed"""
     command: Optional[V2_CMD] = _anoinit(default=None)
@@ -295,7 +296,8 @@ class ProxyData:
     If not an empty string, contains the error encountered when parsing
     """
     _tlv: Optional[ProxyTLV] = _anoinit(default=None)
-
+    # pytype: enable=annotation-type-mismatch
+    
     @property
     def valid(self) -> bool:
         return not (self.error or self.version is None or self.protocol is None)
@@ -316,7 +318,7 @@ class ProxyData:
         :param log_prefix: If True, add "PROXY error:" prefix to log message
         """
         if log_prefix:
-            log.warning(f"PROXY error: {error_msg}")
+            log.warning("PROXY error: %s", error_msg)
         else:
             log.warning(error_msg)
         self.error = error_msg